requirements.txt
Jinja2==2.6
Markdown==2.2.1
pyScss==1.1.4
wsgiref==0.1.2
